Refactor syntax
