--- modules/itc/src/main/scala/lucuma/itc/search/syntax/HIIRegion.scala
@@ -1,25 +1,0 @@
-// Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
-// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
-
-package lucuma.itc.search.syntax
-
-import lucuma.core.enums.HIIRegionSpectrum
-
-/**
- * Syntax extensions for missing ocs2Tag items
- */
-final class HIIRegionSpectrumOps(val self: HIIRegionSpectrum) extends AnyVal {
-
-  def ocs2Tag: String =
-    self match {
-      case HIIRegionSpectrum.OrionNebula => "Orion-nebula"
-    }
-
-}
-
-trait ToHIIRegionSpectrumOps {
-  implicit def toHIIRegionSpectrumOps(self: HIIRegionSpectrum): HIIRegionSpectrumOps =
-    new HIIRegionSpectrumOps(self)
-}
-
-object hiiregionspectrum extends ToHIIRegionSpectrumOps
