+++ modules/itc/src/main/scala/lucuma/itc/search/syntax/package.scala
@@ -1,24 +1,20 @@
 // Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package lucuma.itc.search
+package lucuma.itc.search.syntax
 
-package object syntax {
-  object all
-      extends ToGmosNorthFilterOps
-      with ToGmosNorthGratingOps
-      with ToGmosNorthFpuOps
-      with ToGmosSouthFilterOps
-      with ToGmosSouthGratingOps
-      with ToGmosSouthFpuOps
-      with lucuma.itc.legacy.syntax.ConditionsSyntax
-
-  object sed
-      extends ToStellarLibrarySpectrumOps
-      with ToGalaxySpectrumOps
-      with ToPlanetSpectrumOps
-      with ToHIIRegionSpectrumOps
-      with ToPlanetaryNebulaSpectrumOps
-      with ToQuasarSpectrumOps
-      with ToCoolStarModelOps
-}
+object all
+    extends ToGmosNorthFilterOps
+    with ToGmosNorthGratingOps
+    with ToGmosNorthFpuOps
+    with ToGmosSouthFilterOps
+    with ToGmosSouthGratingOps
+    with ToGmosSouthFpuOps
+    with lucuma.itc.legacy.syntax.ConditionsSyntax
+    with lucuma.itc.legacy.syntax.CoolStarModelSyntax
+    with lucuma.itc.legacy.syntax.StellarLibrarySpectrumSyntax
+    with lucuma.itc.legacy.syntax.GalaxySpectrumSyntax
+    with lucuma.itc.legacy.syntax.PlanetSpectrumSyntax
+    with lucuma.itc.legacy.syntax.HIIRegionSpectrumSyntax
+    with lucuma.itc.legacy.syntax.PlanetaryNebulaSpectrumSyntax
+    with lucuma.itc.legacy.syntax.QuasarSpectrumSyntax
